Completed the implementation of register route
