+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "bcryptjs": "^2.4.3",
     "mongodb": "^6.3.0",
     "mongoose": "^8.0.3",
     "next": "14.0.3",
